"Moved pollModel to db folder and deleted widgets.sql"
